Sets Surf Dirty When Opening Generic XSec GUI for W/H Deactivation
This is added so that the width parm for wings and both width and height
parms for props are deactivated when the GUI is opened. Without this
there is a delay between the GUI sliders being deactivated.

(cherry picked from commit 8cbc1381c749b55da29608a6d01b970182e938dc)