Fixes Snap To Update Issue by Changing to Final Full Update
This was found by Brandon Litherland and was causing the Parm
associated with the minimum distance to be found and updated but the
reported minimum distance, collision detection, and Geom surface
to not update. The final update is changed to a full update.
However, if the value of the Parm is not changing the dirty flags
will not be set. Therefore, a check is added to trigger a Parm
change and set the dirty flags if the Parm value was not changed.