Adds Option to Skip API Unit Tests (i.e. Custom Geom Functions)
A bool is added to identify if a function should be skipped when
generating the API unit test script. This allows functions that can't
be run from *.vspscripts, like those for Custom Geoms to be ignored.
Mapping of comments to functions is improved by mapping the entire
struct in one variable instead of multiple for each field in the struct.