Fixes Issue #181: Adds Missing Support for Prop Conformal Components
PropGeom was missing a virtual OffsetXSecs function, so one is added to
apply an offset to the diameter and first xsec location. The location of
the XSecs and T Parm values are adjusted to maintain the same relative
spacing. The spanwise chord/r values are adjusted as well to apply the
offset in the chordwise direction.